backend/lct_backend2/requirements.txt
# Core FastApi
fastapi
uvicorn[standard]
aio-pika
pydantic
pydantic-settings
httpx

# PyTorch с CUDA 12.4 для RTX 4070 Super
# Установка: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu124
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.5.1+cu124
torchvision==0.20.1+cu124

# ML и обработка изображений
antlr4-python3-runtime==4.9.3
appnope==0.1.4
asttokens==3.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
filelock==3.19.1
fonttools==4.60.0
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.1
hydra-core==1.3.2
idna==3.10
iopath==0.1.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
polars==1.33.1
portalocker==3.2.0
prompt_toolkit==3.0.52
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==27.1.0
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
git+https://github.com/facebookresearch/sam2.git@2b90b9f5ceec907a1c18123530e92e794ad901a4#egg=SAM_2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
tokenizers==0.22.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.2
typing_extensions==4.15.0
ultralytics==8.3.203
ultralytics-thop==2.0.17
urllib3==2.5.0
wcwidth==0.2.14